requirements.txt
django-select2==8.4.0